Editor: Lift out the Layout selector into its own file
